Reapply: Makefile changes to get mclinker to build again.
Turns out, bison is pretty inflexible when a %includes file.h is in the
.yy file: it will always generate the code in the current directory. Since
the current directory in an Android build is not out/ we cannot assume we
have write access to that directory.

Therefore, the only option I could find was to change the ScriptParser.yy
file. Not ideal, since this is the only patch on upstream.

Change-Id: I9ee7f56d4f6900e1b566f4a36b31dddc8cb0848e